Cope with recent change to composer that prompts about plugins
Composer 2.2 introduced a change to security settings that means that
it now prompts if a plugin wants to do stuff.

Drupal's recommended-project uses such plugins, and its composer.json
hasn't yet been updated to allow the plugins it needs.

So for now, pass in the "-n" command-line option to composer when
creating the project, to skip the prompts. But note that this
is supposed to stop working in July 2022, so we need to revisit this
later.

After creation of the project, add the permissions for the plugins
used.

Also allow the plugin used by composer-patches. We will probably have
to continue to do this ourselves, as it isn't used out-of-the-box by
Drupal.